Legacy ortools behind API : correction due to review
